+++ src/definition.ts
@@ -1,11 +1,11 @@
 import * as vscode from 'vscode';
 
 export interface ICheckResult {
-	file: string;
-	line: number;
-	column: number;
-	msg: string;
-	severity: string;
+    file: string;
+    line: number;
+    column: number;
+    msg: string;
+    severity: string;
 }
 
 export class TextSelection {
@@ -91,8 +91,16 @@ export class ABLMethod {
     }
 }
 
+export class ABLClassMethod extends ABLMethod { }
+
 export class ABLParameter extends ABLVariable {
     direction: ABL_PARAM_DIRECTION;
+}
+
+export interface ABLClass {
+    className: string;
+    filePath?: string;
+    methods?: ABLClassMethod[];
 }
 
 export class ABLInclude {